Add SigC::Signal#<>::slot(). Comment out make_slot() work-around. Remove
2003-10-30  Martin Schulze  <teebaum@cvs.gnome.org>

	* sigc++/signal.h.m4: Add SigC::Signal#<>::slot().
	* sigc++/slot.h.m4: Comment out make_slot() work-around.
	* sigc++/adaptors/bind.h.m4: Remove unnecessary brackets
	in template argument lists. They are confusing MSVC.
	* sigc++/adaptors/*.h.m4, sigc++/adaptors/lambda/*.h.m4:
	Use LIBSIGC_TEMPLATE_PREFIX in explicit function
	template instantiations.
	* sigc++/tests/test_*.cc:
	- Include <string> where std::string is used.
	- Use double instead of float.

+++ tests/test_accumulated.cc
@@ -23,8 +23,8 @@ struct arithmetic_mean_accumulator
     }
 };
 
-int foo(int i)   { std::cout << "foo: " << 3*i+1 << std::endl; return 3*i+1;}
-int bar(float i) { std::cout << "bar: " << 5*(int)i-3 << std::endl; return 5*(int)i-3;}
+int foo(int i)    { std::cout << "foo: " << 3*i+1 << std::endl; return 3*i+1;}
+int bar(double i) { std::cout << "bar: " << 5*(int)i-3 << std::endl; return 5*(int)i-3;}
 
 struct A : public sigc::trackable 
 {
